requirements.txt
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.25.0
ta>=0.11.0
fredapi>=0.5.1
rich>=13.7.0
pyyaml>=6.0.1
schedule>=1.2.1
requests>=2.31.0
anthropic>=0.39.0
